Added scoring; fixed Angry Bird duplication upon collision
The AngryBird and Pig are now Actionable objects, which indicates to the collision handler in GamePlatform that they can handle their own collision events.

The AngryBird collision code only runs on the first collision event to avoid spawning duplicate AngryBirds when it bounces, for example.

Bird score does not update at the moment, until we decide how that should work.